+++ hypnotoad/core/mesh.py
@@ -3534,6 +3534,15 @@ class BoutMesh(Mesh):
             f.write("curvature_type", self.user_options.curvature_type)
             f.write("Bt_axis", self.equilibrium.Bt_axis)
 
+            if hasattr(self.equilibrium, "psi_axis"):
+                f.write("psi_axis", self.equilibrium.psi_axis)
+            if hasattr(self.equilibrium, "psi_bdry"):
+                f.write("psi_bdry", self.equilibrium.psi_bdry)
+            if hasattr(self.equilibrium, "psi_axis_gfile"):
+                f.write("psi_axis_gfile", self.equilibrium.psi_axis_gfile)
+            if hasattr(self.equilibrium, "psi_bdry_gfile"):
+                f.write("psi_bdry_gfile", self.equilibrium.psi_bdry_gfile)
+
             # write the 2d fields
             for name in self.fields_to_output:
                 self.writeArray(name, self.__dict__[name], f)
@@ -3726,6 +3735,7 @@ class BoutMesh(Mesh):
                 f.write_file_attribute(
                     "hypnotoad_geqdsk_filename", self.equilibrium.geqdsk_filename
                 )
+
             if hasattr(self.equilibrium, "geqdsk_input"):
                 # If grid was created from a geqdsk file, save the file contents
                 #